IICS - SOURCE TO TARGET
Login to INFORMATICA INTELLIGENT XLOUD SERVICE (IICS) - Click DATA INTEGRATION.
Create a new project/folder in EXPLORE - Give name and description.
Go to the respected project folder .
Click new - Click Mappings - Create a mapping to deploy data integration logic that is processed on the secure agent.
Opens a window where we perform data transformation
Change the name of the Mapping
Click on Source - Go to Source in the properties - Select the connection - Source type Single - Select the object/table.
Click on Target - Go to Target in the properties - Select the connection - Source type Single - Create the object/table.
Save and Run - Go to Jobs and view the output.