IICS - SORTING
Go to the respective project folder .
Click new - Click Mappings - Create a mapping to deploy data integration logic that is processed on the secure agent.
When the Mapping is created it will direct to the mapping page where we perform data transformation.
Change the name of the Mapping
Click on Source - Go to Source in the properties - Select the connection - Source type Single - Select the object/table.
There are many Transformation available in the Mapping - Sorting, Joiner, Router etc.
In this project the sorting transformation is used to sort the employee name in the alphabetic wise (A-Z).
Drag and Drop the Sorter Transformation.
Click on Sorter - Go to Sort in the properties - Add Sort Condition - Select the Field to the sorted and Select the Sort Order.
Click on Target - Go to Target in the properties - Select the connection - Source type Single - Create the object/table.
Save and Run.