IICS - EXPRESSION
Go to the respective project folder .
Click new - Click Mappings - Create a mapping to deploy data integration logic that is processed on the secure agent.
Change the Mapping name.
Click on Source - Go to Source in the properties - Select the connection - Source type Single - Select the object/table.
When the Mapping is created it will direct to the mapping page where we perform data transformation.
There are many Transformation available in the Mapping - Sorting, Joiner, Router etc.
In this project the Expression is used to Concatenate the employee First Name and Last name - Adding a new Column called Full Name.
Drag and Drop the Expression Transformation.
Click on Expression - Go to Expression in the properties - Add Expression - It insists to type a name to the new column - Click on Expression field - Drag and Drop the columns and perform in the expression in the Expression Box.
Click on Target - Go to Target in the properties - Select the connection - Source type Single - Create the object/table.
Save and Run.