IICS - JOINER
Go to the respective project folder .
Click new - Click Mappings - Create a mapping to deploy data integration logic that is processed on the secure agent.
Change the Mapping name.
Add Two Sources.
Click on Source - Go to Source in the properties - Select the connection - Source type Single - Select the object/table.
When the Mapping is created it will direct to the mapping page where we perform data transformation.
There are many Transformation available in the Mapping - Sorting, Joiner, Router etc.
In this project the Joiner is used to perform Full outer Join to Join all the table from two tables into one table.
Drag and Drop the Joiner Transformation.
Click on Joiner - Go to Joiner Condition in the properties - Mention the Join type as Full Outer - Join Condition as Simple -Add Join Condition as ID of one source equal to the ID of another source
Click on Target - Go to Target in the properties - Select the connection - Source type Single - Create the object/table.
Save and Run