IICS - AGGREGATE
Go to the respective project folder .
Click new - Click Mappings - Create a mapping to deploy data integration logic that is processed on the secure agent.
Change the Mapping name.
Click on Source - Go to Source in the properties - Select the connection - Source type Single - Select the object/table (Health Insurance).
There are many Transformation available in the Mapping - Sorting, Joiner, Router etc.
In this project the Aggregate is used to perform Max of AGE Group By GENDER.
Drag and Drop the Aggregate Transformation.
Click on Aggregate - Go to Group by section and add the field to be grouped - Go to Aggregate section and add the condition - MAX(AGE).
Click on Target - Go to Target in the properties - Select the connection - Source type Single - Create the object/table.
Save and Run.